move messages to different file
